# perf: lazyop as dataclass