--- a/test/external/external_test_allocator_on_models.py
+++ b/test/external/external_test_allocator_on_models.py
@@ -2,11 +2,12 @@
 import unittest, gc
 import numpy as np
 from tinygrad.tensor import Tensor
-from tinygrad.nn.state import get_parameters, get_state_dict
-from tinygrad.ops import GlobalCounters, LazyOp, LoadOps
+from tinygrad.nn.state import get_state_dict
+from tinygrad.ops import GlobalCounters
 from tinygrad.runtime.lib import RawBuffer, LRUAllocator
 from tinygrad.helpers import dtypes, prod
 from tinygrad.ops import Device
+from test.helpers import derandomize_model
 
 from examples.llama import Transformer
 
@@ -85,20 +86,6 @@ def check_gc():
     gc.collect() # Need to collect Tensors.
     from extra.introspection import print_objects
     assert print_objects() == 0
-
-# for speed
-def derandomize(x):
-  if isinstance(x, LazyOp):
-    if x.op == LoadOps.RAND: x.op = LoadOps.EMPTY
-    x.src = tuple([derandomize(s) for s in x.src])
-  else:
-    x.op = derandomize(x.op)
-  return x
-
-def derandomize_model(model):
-  for p in get_parameters(model):
-    p.lazydata = derandomize(p.lazydata)
-    p.realize()
 
 class TestAllocators(unittest.TestCase):
   @unittest.skipUnless(Device.DEFAULT == "GPU", "Not Implemented")