--- a/tinygrad/ops.py
+++ b/tinygrad/ops.py
@@ -7,7 +7,7 @@ from typing import TYPE_CHECKING, Union, Type, Tuple, Any, List, Optional, Dict,
 from tinygrad.helpers import ansilen, prod, DEBUG, getenv, GlobalCounters, DType, colored
 from tinygrad.runtime.lib import RawBuffer
 from tinygrad.shape.symbolic import Variable, sym_infer
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 
 # these are the llops your accelerator must implement, along with toCpu
 # the Enum class doesn't work with mypy, this is static. sorry it's ugly
@@ -48,21 +48,20 @@ class ScheduleItem:
   inputs: Tuple[LazyBuffer, ...]
   var_vals: Dict[Variable, int]
 
-@dataclass
+@dataclass(frozen=True)
 class LazyOp:
   op: Op
   src: Tuple[Union[LazyOp, LazyBuffer], ...]
   arg: Any = None
-  buffers: Tuple[LazyBuffer, ...] = field(init=False)
-  def __post_init__(self):
-    self.buffers = ()
+
+  @property
+  def buffers(self):
+    buffers = ()
     try:
-      for x in self.src: self.buffers += x.buffers
-    except AttributeError: self.buffers = ()
-    # NOTE: the linearizer's key function maps the buffers to ints, and LOCAL_BUFFER is used. we don't care about buffers in these cases
-  def __repr__(self): return f"LazyOp(op={self.op}, src={self.src}, arg={self.arg})"
-  def __eq__(self, __value: object) -> bool: return isinstance(__value, LazyOp) and self.op is __value.op and self.src == __value.src and self.arg == __value.arg
-  def __hash__(self) -> int: return hash((self.op, self.src, self.arg))
+      for x in self.src: buffers += x.buffers
+    except AttributeError: return () # NOTE: the linearizer's key function maps the buffers to ints, and LOCAL_BUFFER is used. we don't care about buffers in these cases
+    return buffers
+
   @property
   def key(self): return (self.op, tuple(map(lambda x: getattr(x, "key", x), self.src)), getattr(self.arg, "key", self.arg))