before
```
codegen         mean runtime:  189.89ms, runs:   194.70,  239.41,  180.85,  170.33,  207.72,  178.27,  179.02,  183.95,  183.69,  181.00
methodcache     mean runtime:  176.97ms, runs:   170.78,  224.18,  163.98,  163.25,  164.45,  189.46,  193.26,  167.76,  166.27,  166.29
profile         mean runtime:  790.07ms, runs:   769.91,  794.76,  799.09,  753.40,  847.84,  762.01,  795.60,  804.17,  770.68,  803.29
```


after
```
codegen         mean runtime:  182.52ms, runs:   182.84,  250.48,  175.73,  165.99,  176.54,  173.80,  172.70,  179.16,  173.53,  174.45
methodcache     mean runtime:  165.79ms, runs:   159.17,  220.91,  158.91,  177.57,  155.09,  155.78,  160.16,  153.85,  154.83,  161.62
profile         mean runtime:  785.40ms, runs:   758.52,  798.52,  801.80,  765.71,  780.81,  781.45,  806.35,  778.69,  778.43,  803.76
````