@tinyb0t 2023-08-21
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/ops.py                            7              12
------------------------------------------------------------
lines added in the tinygrad folder: -5
```