@geohot 2023-08-21
Err, this is confusing. What's a dataclass? Can this be a NamedTuple, I understand NamedTuple