@roelofvandijk 2023-08-22
They are similar in that they offer quite some convenience methods out-of-the box, such as an `__init__` to store the args in members, but a dataclass need not be immutable (a NamedTuple always is). `@dataclass(frozen=True)` is immutable (and hashable). 

Here is a pretty good comparison:
https://death.andgravity.com/namedtuples

Let me see if this can become a `NamedTuple`.