@roelofvandijk 2023-08-22
Ah yes, two blockers:
* NamedTuple cannot have a weakref.
* this line modifies one of the attributes: https://github.com/tinygrad/tinygrad/blob/master/tinygrad/codegen/linearizer.py#L518.
I think a dataclass is really the nicest choice here.