@roelofvandijk 2023-08-22
In general: there is quite a lot to be gained in terms of speed in the object initialization of Tensors, Functions, Buffers and Ops.