@wozeparrot 2023-08-28
So I do kinda support this, `LazyOp` does make sense as a dataclass because of how it's initialized, and generally it is cleaner for me to read at least. But how much performance does having it be a dataclass give us? I thought dataclass was just syntax sugar around normal classes, or is it closer to a namedtuple during execution?