@roelofvandijk 2023-10-06
Frozen data classes are very close, if not faster than namedtuples (in terms of init, property access, methods etc.). Also hashable and immutable by default.