@roelofvandijk 2023-10-11
1) Frozen dataclass
2) Moved buffers to property to allow freezing. This also makes the hash, repr and eq functions the natural ones.

Did not focus on speed but on readability here.